+++ NEWS.md
@@ -2,6 +2,7 @@
 
 * Log level by default :notice is removed. Use :info. 
 * Added ActArray#power? method
+* Added #beams and #stored attributes for Gripper
 
 ## 2012-03-06 0.4.1